pyproject.toml
[project]
name = "WordleWise-backend"
version = "0.1.0"
description = "Backend for WordleWise"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
]
